+++ anyblok/config.py
@@ -89,6 +89,7 @@ def nargs_type(key, nargs, cast):
 class AnyBlokActionsContainer:
 
     def add_argument(self, *args, **kwargs):
+        """Overload the method to add the entry in the configuration dict"""
         default = kwargs.pop('default', None)
         nargs = kwargs.get('nargs', None)
         type = kwargs.get('type')
@@ -108,6 +109,15 @@ class AnyBlokActionsContainer:
 
         Configuration.add_argument(dest, default, type=type)
         return arg
+
+    def set_defaults(self, **kwargs):
+        """Overload the method to update the entry in the configuration dict"""
+        super(AnyBlokActionsContainer, self).set_defaults(**kwargs)
+        for dest, default in kwargs.items():
+            if not Configuration.has(dest):
+                raise KeyError('Unexisting option %s' % dest)
+
+            Configuration.set(dest, default)
 
     def add_argument_group(self, *args, **kwargs):
         group = AnyBlokArgumentGroup(self, *args, **kwargs)
@@ -712,7 +722,6 @@ def add_install_bloks(parser):
     parser.add_argument('--install-bloks', nargs="+", help="blok to install")
     parser.add_argument('--install-all-bloks', action='store_true')
     parser.add_argument('--test-blok-at-install', action='store_true')
-    parser.set_defaults(test_blok=False)
 
 
 @Configuration.add('uninstall-bloks')