+++ doc/CHANGES.rst
@@ -14,6 +14,7 @@ CHANGELOG
 * [FIX] test event
 * [FIX] test blok
 * [FIX] mapper with None parameter
+* [FIX] add set_defaults in parser to update configuration dict
 
 0.8.5 (2016-06-20)
 ------------------